dev/backend/requirements.txt
Flask
Flask-CORS
mysql-connector
